Flytt målgruppetyper i egen fil
